lkl: add support for Microsoft C compiler
Add support for compiling projects that use LKL with the Microsoft C
compiler. This does not yet permit compiling the whole LKL with this
compiler, but external C that includes linux kernel files will now
compile.

This permits compiling applications that use LKL with the Microsoft C
compiler, even though the LKL library itself must be build with mingw
or another supported compiler.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@cs.pub.ro>